backend/requirements.txt
beautifulsoup4==4.12.2
Django==4.2.4
djangorestframework==3.14.0
django-cors-headers==4.2.0
djoser==2.2.0
Requests==2.31.0
lxml==4.9.3
python-dotenv==1.0.1
google-generativeai==0.8.2